Extract flask code into separate class
