+++ backend/data.py
@@ -15,7 +15,7 @@ class SupabaseManager():
         key = os.getenv('SUPABASE_KEY')
         return create_client(url, key)
 
-    def add_message_response(self, question: str, answer: str) -> int:
+    def add_question_answer(self, question: str, answer: str) -> int:
         """Add a question and answer to the database.
 
         Args:
@@ -29,7 +29,7 @@ class SupabaseManager():
             {"question": question, "answer": answer}).execute()
         return data.data[0]["id"]
 
-    def get_messages_answers(self) -> List[Dict[str, Union[int, str]]]:
+    def get_question_answers(self) -> List[Dict[str, Union[int, str]]]:
         """Get all QA.
 
         Returns:
